Add .gitignore file and update project configurations
A new .gitignore file is added to ignore unnecessary files and folders that get generated during build or runtime. Updates are also made to the project's AssemblyInfo and configurations. The NuGet restore properties related to audit are removed, and the documentation paths for Debug and Release configurations are specified in the project file.